- Drones are rotated to an upright position - The player is not able to procede to the next level untill they have filled the energy bar 100% - Lorebox Prefab and script - There is a continue button on the name entry screen - Scanned enemies no longer appear on minimap when disconnection from the Sensors Subsystem - Load Screen no longer appears when pausing - Game now autosaves on level load - Esxisting player inventory is shown on the store screen - Sorashiima Skeleton Layout

+++ Assets/Scripts/UI/ProbeStoreUpgrade.cs
@@ -21,7 +21,7 @@ public class ProbeStoreUpgrade : MonoBehaviour, IPointerClickHandler, IPointerEn
 
     public void OnPointerClick(PointerEventData eventData)
     {
-        if (purchaseing.UpgradeProbe((int)Type))
+        if (purchaseing.UpgradeProbe(Type))
         {
             Checkmark.gameObject.SetActive(true);
             OnPointerEnter(eventData);
